- Incorrect preservation of tracks in the history - Added to crayon - Added to the menu Tutorial tutorial - Added titles everywhere where their nebylo - Localization of everything in the world - Added Info in English - Removed personal data with settings and from the Tutorial - I tested and made a bugfix (mostly UI bugs)

+++ MetTracker/Models/Categories/MTActivity.swift
@@ -21,7 +21,7 @@ class MTActivity: MTSearchObject {
         
         self.activityId = data.object(forKey: "ACTIVITY ID") as! Int
         self.ainsworthCode = data.object(forKey: "AINSWORTH CODE") as! Int
-        self.mets = data.object(forKey: "METS") as? Float ?? 0
+        self.mets = Float(data.object(forKey: "METS") as? Double ?? 0)
         self.name = data.object(forKey: LS("activity_name")) as! String
         self.orderNumber = data.object(forKey: "ORDER NUMBER") as! Int
         self.subcategory = data.object(forKey: "SUBCATEGORY") as! String
